requirements.txt
cryptography>=2.5
django>=2.2.0
django-object-tool>=0.0.10
jsonfield2>=3.0.3
wechatpy>=1.8.3,<2.0.0